+++ scripts/JSRootPainter.hist3d.js
@@ -91,7 +91,7 @@
       }
 
       this.usesvg = JSROOT.BatchMode; // SVG used in batch mode
-      if (this.usesvg) this.usesvgimg = true;
+      //if (this.usesvg) this.usesvgimg = JSROOT.gStyle.ImageSVG;  // use svg images to insert graphics
 
       var sz = this.size_for_3d(this.usesvg ? 3 : undefined);
 
@@ -132,48 +132,11 @@
       this.camera.lookAt(lookat);
       this.scene.add( this.camera );
 
-      var dom_elem = null;
+      var res = JSROOT.Painter.Create3DRenderer(this.scene_width, this.scene_height, this.usesvg);
 
-      if (this.usesvg) {
-         console.log('Creating svg', sz.can3d, sz.width + 'x' + sz.height, 'webgl', JSROOT.Painter.TestWebGL());
-
-         if (this.usesvgimg) {
-            this.webgl = JSROOT.Painter.TestWebGL();
-
-            var node_canvas = require('canvas');
-            var canv = new node_canvas(sz.width, sz.height);
-            canv.style = {};
-
-            console.log('canvas property', canv.width, canv.height, 'style', canv.style);
-
-            this.renderer = this.webgl ? new THREE.WebGLRenderer({ canvas: canv, antialias: true, alpha: true }) :
-                                         new THREE.CanvasRenderer({ canvas: canv, antialias: true, alpha: true });
-         } else {
-            // this.renderer = new THREE.SVGRenderer({ precision: 0, astext: true });
-            this.renderer = THREE.CreateSVGRenderer(false, 0, document);
-         }
-
-         if (this.usesvgimg || (this.renderer.makeOuterHTML !== undefined)) {
-            // this is indication of new three.js functionality
-            if (!JSROOT.svg_workaround) JSROOT.svg_workaround = [];
-            this.renderer.workaround_id = JSROOT.svg_workaround.length;
-            JSROOT.svg_workaround[this.renderer.workaround_id] = "<svg></svg>"; // dummy, need to be replaced
-
-            // replace DOM element in renderer
-            dom_elem = document.createElementNS( 'http://www.w3.org/2000/svg', 'path');
-            dom_elem.setAttribute('jsroot_svg_workaround', this.renderer.workaround_id);
-         }
-      } else {
-         this.webgl = JSROOT.Painter.TestWebGL();
-         this.renderer = this.webgl ? new THREE.WebGLRenderer({ antialias: true, alpha: true }) :
-                                      new THREE.CanvasRenderer({ antialias: true, alpha: true });
-      }
-
-      //renderer.setClearColor(0xffffff, 1);
-      // renderer.setClearColor(0x0, 0);
-      this.renderer.setSize(this.scene_width, this.scene_height);
-
-      this.add_3d_canvas(sz, dom_elem || this.renderer.domElement);
+      this.renderer = res.renderer;
+      this.webgl = res.webgl;
+      this.add_3d_canvas(sz, res.dom);
 
       this.first_render_tm = 0;
       this.enable_highlight = false;
@@ -307,17 +270,7 @@
             console.log('First render tm = ' + this.first_render_tm);
          }
 
-         // when using SVGrenderer producing text output, provide result
-         if (this.renderer.workaround_id !== undefined) {
-            if (this.usesvgimg) {
-               var canvas = this.renderer.domElement;
-               var dataUrl = canvas.toDataURL("image/png");
-               var svg = '<image width="' + canvas.width + '" height="' + canvas.height + '" xlink:href="' + dataUrl + '"></image>';
-               JSROOT.svg_workaround[this.renderer.workaround_id] = svg;
-            } else {
-               JSROOT.svg_workaround[this.renderer.workaround_id] = this.renderer.makeOuterHTML();
-            }
-         }
+         JSROOT.Painter.AfterRender3D(this.renderer);
 
          return;
       }